+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - cmd/script: allow executing script with arguments.
 - cmd/hook: install hook as pre-push rather than pre-commit.
+- cmd/script: prevent from adding blank script.
 
 ## [0.7.1] - 2021-02-11